Chapter 12 Exception Handling
This was a long, drawn out exercise.  My objective was to practice exception handling.  I was handling exceptions that can be thrown when a user inputs values into the program to divide simple integers (Scanner entry determined numerator and denominator values).  While looping keeps the program moving and robust, even if a user enters a non-integer or evaluates an illegal expression (divide by 0).  It was a struggle to get the while loop and exceptions to work in harmony.  Remembering to clear the keyboard buffer via Scanner.nextLine();